lambda_message/requirements.txt
slackclient>=2.0.0
slackeventsapi>=2.1.0
Flask>=1.1.2
#dateparser>=0.7.6
requests>=2.24.0
